ec: move scrub, replication check, and shard unmount engines into weed/ec
