+++ weed/ec/ec_check_replication_test.go
@@ -1,4 +1,4 @@
-package shell
+package ec
 
 import (
 	"bytes"
