+++ weed/shell/command_ec_shard_unmount.go
@@ -1,76 +1,19 @@
 package shell
 
 import (
-	"context"
 	"flag"
 	"fmt"
 	"io"
 	"math"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
 
-	"github.com/seaweedfs/seaweedfs/weed/operation"
-	"github.com/seaweedfs/seaweedfs/weed/pb"
-	"github.com/seaweedfs/seaweedfs/weed/pb/master_pb"
-	"github.com/seaweedfs/seaweedfs/weed/pb/volume_server_pb"
-	"github.com/seaweedfs/seaweedfs/weed/storage/erasure_coding"
+	"github.com/seaweedfs/seaweedfs/weed/ec"
 )
-
-const (
-	ecShardActionTimeout = 1 * time.Minute
-)
-
-type ecShard struct {
-	ShardID     uint32
-	Collection  string
-	NodeAddress string
-}
-
-func (s *ecShard) String() string {
-	if s.NodeAddress == "" {
-		return fmt.Sprintf("%d", s.ShardID)
-	}
-	return fmt.Sprintf("%d@%s", s.ShardID, s.NodeAddress)
-}
-
-func ecShardsFromString(shards string) ([]*ecShard, error) {
-	res := []*ecShard{}
-
-	for _, s := range strings.Split(shards, ",") {
-		s = strings.TrimSpace(s)
-		if s == "" {
-			return nil, fmt.Errorf("empty shard ID in %q", shards)
-		}
-
-		// optional <shard ID>@<node address> to pick one copy
-		idStr, addr, _ := strings.Cut(s, "@")
-
-		id, err := strconv.Atoi(idStr)
-		if err != nil || id < 0 || id >= erasure_coding.MaxShardCount {
-			return nil, fmt.Errorf("invalid shard ID %q", s)
-		}
-
-		res = append(res, &ecShard{ShardID: uint32(id), NodeAddress: addr})
-	}
-
-	return res, nil
-}
 
 func init() {
 	Commands = append(Commands, &commandEcShardUnmount{})
 }
 
 type commandEcShardUnmount struct {
-	env           *CommandEnv
-	writer        io.Writer
-	topology      *master_pb.TopologyInfo
-	volumeID      uint32
-	delete        bool
-	ignoreInvalid bool
-	apply         bool
-	shards        []*ecShard
 }
 
 func (c *commandEcShardUnmount) Name() string {
@@ -107,7 +50,6 @@ func (c *commandEcShardUnmount) Do(args []string, commandEnv *CommandEnv, writer
 	if handleHelpRequest(c, args, writer) {
 		return nil
 	}
-
 	ecShardUnmountCommand := flag.NewFlagSet(c.Name(), flag.ContinueOnError)
 	volumeID := ecShardUnmountCommand.Uint("volumeId", 0, "volume ID for the shards to process")
 	shardIDsStr := ecShardUnmountCommand.String("shardId", "", "comma-separated EC shard IDs for the volume")
@@ -133,7 +75,7 @@ func (c *commandEcShardUnmount) Do(args []string, commandEnv *CommandEnv, writer
 	if *shardIDsStr == "" {
 		return fmt.Errorf("missing shardId")
 	}
-	shards, err := ecShardsFromString(*shardIDsStr)
+	shards, err := ec.ShardRefsFromString(*shardIDsStr)
 	if err != nil {
 		return err
 	}
@@ -144,136 +86,12 @@ func (c *commandEcShardUnmount) Do(args []string, commandEnv *CommandEnv, writer
 		return err
 	}
 
-	c.env = commandEnv
-	c.writer = writer
-	c.topology = topology
-	c.volumeID = uint32(*volumeID)
-	c.delete = *delete
-	c.ignoreInvalid = *ignoreInvalid
-	c.apply = *apply || *applyAlias
-	c.shards = shards
-
-	return c.doShardsUnmount()
-}
-
-func (c *commandEcShardUnmount) write(format string, a ...any) {
-	fmt.Fprintf(c.writer, format, a...)
-}
-
-func (c *commandEcShardUnmount) liveShardsForVolume() []*ecShard {
-	shards := []*ecShard{}
-
-	for _, dci := range c.topology.GetDataCenterInfos() {
-		for _, ri := range dci.GetRackInfos() {
-			for _, dni := range ri.GetDataNodeInfos() {
-				nodeAddress := dni.GetAddress()
-				for _, di := range dni.GetDiskInfos() {
-					for _, eci := range di.GetEcShardInfos() {
-						if eci.GetId() == c.volumeID {
-							sinfo := erasure_coding.ShardsInfoFromVolumeEcShardInformationMessage(eci)
-							for _, sid := range sinfo.Ids() {
-								shards = append(shards, &ecShard{
-									ShardID:     uint32(sid),
-									Collection:  eci.GetCollection(),
-									NodeAddress: nodeAddress,
-								})
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	sort.SliceStable(shards, func(i, j int) bool { return shards[i].ShardID < shards[j].ShardID })
-	return shards
-}
-
-func (c *commandEcShardUnmount) printShards(ss []*ecShard) {
-	for _, s := range ss {
-		c.write("\t%v\n", s)
-	}
-	c.write("\n")
-}
-
-func (c *commandEcShardUnmount) doShardsUnmount() error {
-	liveShards := c.liveShardsForVolume()
-	c.write("Live shard topology for volume ID %d (%d shards):\n", c.volumeID, len(liveShards))
-	c.printShards(liveShards)
-
-	// resolve target shards against the live topology
-	targetShards := []*ecShard{}
-	for _, ps := range c.shards {
-		var result *ecShard
-		for _, ts := range liveShards {
-			if ts.ShardID == ps.ShardID {
-				if ps.NodeAddress == "" || ps.NodeAddress == ts.NodeAddress {
-					if result != nil {
-						return fmt.Errorf("shard %v is ambiguous", ps)
-					}
-					result = ts
-				}
-			}
-		}
-		if result == nil {
-			if !c.ignoreInvalid {
-				return fmt.Errorf("shard %v is invalid", ps)
-			}
-			c.write("!!! ignoring invalid shard %v\n", ps)
-		} else {
-			targetShards = append(targetShards, result)
-		}
-	}
-	if len(targetShards) == 0 {
-		return fmt.Errorf("got no shards to process")
-	}
-
-	mode := "unmount"
-	if c.delete {
-		mode = "unmount + delete"
-	}
-	c.write("Will %s %d shard(s):\n", mode, len(targetShards))
-	c.printShards(targetShards)
-
-	if !c.apply {
-		c.write("Not proceeding in dry-run mode\n")
-		return nil
-	}
-
-	for _, s := range targetShards {
-		if err := c.unmountShard(s); err != nil {
-			return err
-		}
-	}
-
-	c.write("\nAll done!\n")
-	return nil
-}
-
-func (c *commandEcShardUnmount) unmountShard(s *ecShard) error {
-	ctx, cancel := context.WithTimeout(context.Background(), ecShardActionTimeout)
-	defer cancel()
-
-	return operation.WithVolumeServerClient(false, pb.ServerAddress(s.NodeAddress), c.env.option.GrpcDialOption, func(vsc volume_server_pb.VolumeServerClient) error {
-		c.write("Unmounting shard %v for volume ID %d...\n", s, c.volumeID)
-		if _, err := vsc.VolumeEcShardsUnmount(ctx, &volume_server_pb.VolumeEcShardsUnmountRequest{
-			VolumeId: c.volumeID,
-			ShardIds: []uint32{s.ShardID},
-		}); err != nil {
-			return err
-		}
-
-		if c.delete {
-			c.write("Deleting shard %v for volume ID %d...\n", s, c.volumeID)
-			if _, err := vsc.VolumeEcShardsDelete(ctx, &volume_server_pb.VolumeEcShardsDeleteRequest{
-				VolumeId:   c.volumeID,
-				Collection: s.Collection,
-				ShardIds:   []uint32{s.ShardID},
-			}); err != nil {
-				return err
-			}
-		}
-
-		return nil
+	return ec.UnmountShards(commandEnv.ecEnv(), writer, ec.ShardUnmountRequest{
+		Topology:      topology,
+		VolumeID:      uint32(*volumeID),
+		Shards:        shards,
+		Delete:        *delete,
+		IgnoreInvalid: *ignoreInvalid,
+		Apply:         *apply || *applyAlias,
 	})
 }
